add section menu for the project
