+++ src/index.js
@@ -1,8 +1,8 @@
 import 'bootstrap';
 import navbar from './tab/navbar';
 import sectionInfo from './tab/sectionInfo';
+import sectionMenu from './tab/sectionMenu';
 import './scss/app.scss';
-
 
 function componentHeader() {
   var element = document.createElement('header');
@@ -10,7 +10,6 @@ function componentHeader() {
   return element;
 }
 
-
-
 document.body.appendChild(componentHeader());
 document.body.appendChild(sectionInfo);
+document.body.appendChild(sectionMenu);
