+++ src/tab/sectionInfo.js
@@ -2,7 +2,7 @@ import bg from '../img/bgTop.jpg';
 
 const sectionInfo = document.createElement('section');
 sectionInfo.className ='wrapper-info min-mw-100  min-vh-100 p-5';
-sectionInfo.id = 'section-info';
+sectionInfo.setAttribute('id','section-info');
 
 const h1 = document.createElement('h1');
 h1.appendChild(document.createTextNode('El Truco Restaurant'));
